data cleaning script created
