+++ .idea/.name
@@ -1,1 +1,1 @@
-glassdoor_scraperV2.py
+glassdoor_scraper.py
